move bearer stategy into strategies folder
